Move DatabaseModule to separate file
