+++ app/src/main/java/com/bhaktaprogram/notehelper/main/AppComponent.kt
@@ -1,12 +1,8 @@
 package com.bhaktaprogram.notehelper.main
 
 import android.content.Context
-import androidx.room.Room
-import com.bhaktaprogram.notehelper.db.AppDatabase
 import dagger.BindsInstance
 import dagger.Component
-import dagger.Module
-import dagger.Provides
 import javax.inject.Singleton
 
 @Singleton
@@ -22,22 +18,5 @@ interface AppComponent {
         fun setContext(context: Context): Builder
 
         fun build(): AppComponent
-    }
-}
-
-@Module
-class DatabaseModule {
-
-    companion object {
-
-        private const val DATABASE_NAME = "notes.db"
-
-        @Provides
-        @Singleton
-        fun provideDatabase(context: Context) = Room.databaseBuilder(
-            context,
-            AppDatabase::class.java,
-            DATABASE_NAME
-        ).build()
     }
 }
